feat:home manager switch, wip
